+++ Changelog.md
@@ -1,3 +1,7 @@
+#### 2018-03-06:
+ - \#328: Using ${#var} produces incorrect warning "Array use of non-array variable"
+ - \#517: Incorrect warning "Array use of non-array variable"
+
 #### 2018-02-02:
  - \#505: Fixed "Lexer could not match input" error for redirect after here string
  - \#508: Argument values of the mapfile/readarray built-in were parsed as variable names